Handle trigger deployments in exe and improve arg handling

+++ src/solutions/Neo42.DeployNxtApplication/DeployNxtApplication.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Windows.Forms;
 using Microsoft.PowerShell;
 using System.Management.Automation.Language;
-using System.Security.Principal;
+using System.Text.RegularExpressions;
 
 namespace Neo42.DeployNxtApplication
 {
@@ -15,17 +15,21 @@ namespace Neo42.DeployNxtApplication
 	{
 		private static readonly string _deployScript = Path.Combine(Path.GetDirectoryName(Application.ExecutablePath), "Deploy-Application.ps1");
 
-		private static readonly List<string> _deployScriptArguments = [.. Environment.GetCommandLineArgs().Skip(1)];
-
 		private static readonly string _deployDirectory = Path.GetDirectoryName(Application.ExecutablePath);
 
 		private static readonly string _powerShellPath = Environment.Is64BitOperatingSystem && !Environment.Is64BitProcess
 			? $"{Environment.GetFolderPath(Environment.SpecialFolder.Windows)}\\Sysnative\\WindowsPowerShell\\v1.0\\powershell.exe"
 			: $"{Environment.GetFolderPath(Environment.SpecialFolder.System)}\\WindowsPowerShell\\v1.0\\powershell.exe";
 
-		private static readonly bool _isAdmin = new WindowsPrincipal(WindowsIdentity.GetCurrent()).IsInRole(WindowsBuiltInRole.Administrator);
+		private static readonly char[] _stringMarker = ['\'', '"'];
 
-		private static readonly bool _asAdmin = GetPowershellArgument([.. Environment.GetCommandLineArgs().Skip(1)], "DeploymentType", 0)?.EndsWith("UserPart") ?? true;
+		private static readonly List<string> _deployScriptArguments = [.. Environment.GetCommandLineArgs().Skip(1)];
+
+		private static readonly string _deploymentType = GetPowershellArgumentValue(_deployScriptArguments, "DeploymentType", 0) ?? "Install";
+
+		private static readonly bool _asAdmin = !_deploymentType.EndsWith("UserPart", StringComparison.OrdinalIgnoreCase);
+
+		private static readonly bool _isTrigger = _deploymentType.StartsWith("Trigger", StringComparison.OrdinalIgnoreCase);
 
 		private const int ERROR_EXIT_CODE = 60010;
 
@@ -47,13 +51,20 @@ namespace Neo42.DeployNxtApplication
 
 				LoadSettings();
 
+				if (_isTrigger)
+				{
+					var deploymentTypeIndex = GetPowershellArgumentValueIndex(_deployScriptArguments, "DeploymentType", 0);
+					_deployScriptArguments[deploymentTypeIndex] = Regex.Replace(_deployScriptArguments[deploymentTypeIndex], @"\bTrigger", string.Empty, RegexOptions.IgnoreCase);
+				}
+
 				Environment.Exit(
 					ExecutePowerShell(
 						_deployScript,
 						_deployScriptArguments,
 						_executionPolicy,
 						_deployDirectory,
-						_asAdmin
+						_asAdmin,
+						_isTrigger
 					)
 				);
 			}
@@ -99,7 +110,7 @@ namespace Neo42.DeployNxtApplication
 				+ ";exit(@(!$?;gv LASTEXITCODE -va -ea 0)[-1])\"}";
 		}
 
-		private static int ExecutePowerShell(string file, IEnumerable<string> args, ExecutionPolicy execPolicy, string workingDir, bool asAdmin)
+		private static int ExecutePowerShell(string file, IEnumerable<string> args, ExecutionPolicy execPolicy, string workingDir, bool asAdmin, bool detached)
 		{
 			Environment.SetEnvironmentVariable("PSModulePath", null, EnvironmentVariableTarget.Process); // Clear the PSModulePath to avoid module loading issues.
 
@@ -110,19 +121,24 @@ namespace Neo42.DeployNxtApplication
 					Arguments = GetPowerShellArgumentString(file, args, execPolicy),
 					WorkingDirectory = workingDir,
 					WindowStyle = ProcessWindowStyle.Hidden,
-					CreateNoWindow = true,
-					UseShellExecute = asAdmin && !_isAdmin, // UseShellExecute must only be true if elevation is needed. Prevents duplicate UAC and issues with Process.Start() returning null
+					UseShellExecute = true,
 					Verb = asAdmin ? "runas" : null
 				}
 			);
 
-			if (!process.WaitForExit(7200000)) // Wait for up to 120 minutes
+			if (!detached)
 			{
-				process.Kill();
-				throw new TimeoutException($"The PowerShell process '{_powerShellPath}' did not exit within the expected time limit.");
+				if (!process.WaitForExit(7200000)) // Wait for up to 120 minutes
+				{
+					process.Kill();
+					throw new TimeoutException($"The PowerShell process '{_powerShellPath}' did not exit within the expected time limit.");
+				}
+				return process.ExitCode;
 			}
-
-			return process.ExitCode;
+			else
+			{
+				return 0;
+			}
 		}
 
 		private static Hashtable ImportPsDataFile(string path)
@@ -141,36 +157,52 @@ namespace Neo42.DeployNxtApplication
 			return (Hashtable)hashtable.SafeGetValue();
 		}
 
-		private static string? GetPowershellArgument(this List<string> argumentList, string argument, uint? position = null)
+		private static string? GetPowershellArgumentValue(this List<string> argumentList, string argument, uint? position = null)
 		{
-			if (!argumentList.Any())
-			{
-				return null;
-			}
-
-			if (position is not null && argumentList.Count >= position && argumentList.Take((int)position).All(s => !s.StartsWith("-")))
-			{
-				return argumentList[(int)position];
-			}
-
-			var index = argumentList.FindIndex(s => s.Equals("-" + argument, StringComparison.OrdinalIgnoreCase));
+			var index = GetPowershellArgumentValueIndex(argumentList, argument, position);
 			if (index < 0)
 			{
 				return null;
 			}
 
 			var argumentString = argumentList[index];
+
 			if (argumentString.StartsWith("-" + argument + ":", StringComparison.OrdinalIgnoreCase))
 			{
-				return argumentString.Split([':'], 2).Last();
+				argumentString = argumentString.Split([':'], 2).Last();
 			}
 
-			if (argumentList.Count > index + 1 && !argumentList[index + 1].StartsWith("-"))
+			if (argumentString.Length >= 2
+				&& argumentString.First() is char startChar
+				&& argumentString.Last() == startChar
+				&& _stringMarker.Contains(startChar)
+			)
 			{
-				return argumentList[index + 1];
+				argumentString = argumentString.Substring(1, argumentString.Length - 2);
 			}
 
-			return null;
+			return argumentString;
+		}
+
+		private static int GetPowershellArgumentValueIndex(this List<string> argumentList, string argument, uint? position = null)
+		{
+			var index = argumentList.FindIndex(s => s.Equals("-" + argument, StringComparison.OrdinalIgnoreCase) || s.StartsWith("-" + argument + ":", StringComparison.OrdinalIgnoreCase));
+			if (index >= 0)
+			{
+				if (argumentList[index].StartsWith("-" + argument + ":", StringComparison.OrdinalIgnoreCase))
+				{
+					return index;
+				}
+
+				return argumentList.Count > index + 1 && !argumentList[index + 1].StartsWith("-") ? index + 1 : -1;
+			}
+
+			if (position is not null && argumentList.Count > position && argumentList.Take((int)position + 1).All(s => !s.StartsWith("-")))
+			{
+				return (int)position;
+			}
+
+			return -1;
 		}
 	}
 }
